+++ bosesoundtouchapi/__init__.py
@@ -14,6 +14,7 @@ _________________
 from bosesoundtouchapi.soundtouchaudiomodes import SoundTouchAudioModes
 from bosesoundtouchapi.soundtouchclient import SoundTouchClient
 from bosesoundtouchapi.soundtouchdevice import SoundTouchDevice
+from bosesoundtouchapi.soundtouchdiscovery import SoundTouchDiscovery
 from bosesoundtouchapi.soundtoucherror import SoundTouchError
 from bosesoundtouchapi.soundtouchexception import SoundTouchException
 from bosesoundtouchapi.soundtouchhdmicecmodes import SoundTouchHdmiCecModes
@@ -28,6 +29,7 @@ __all__ = [
     'SoundTouchAudioModes',
     'SoundTouchClient',
     'SoundTouchDevice',
+    'SoundTouchDiscovery',
     'SoundTouchError',
     'SoundTouchException',
     'SoundTouchHdmiCecModes',